(368) Added local network config code for Suse.
Networking in SUSE works almost exactly like redhat.
The only difference is the location of the networking scripts.

This has been verified as working on SLES11sp1, but requires a previously configured
eth1 interface (as with redhat)

I have also verified that I can still bring up a localnet on rhel 4,5