Ability to specify shell commands from a string.
This allows a provisioning via shell commands that can be completely
contained within the Vagrantfile, rather than included in an outside
shell script.  Complex shell provisioning might still be better from
and outside file, but this allows for a quick-n-easy way to tweak
the vagrant box image, if necesary, before running other provisioners.