Made some 'Windows specific' updates to the docs
These are some changes that I ran into while I was trying to install
vagrant on my Windows 7 machine.  You might want to check for yourself
if you can reproduce these errors or provide a different solution to
them, but this is how I accomplished the task and I thought I would
share this info in the docs so others don't have to deal with the
headache.

Changes:
--------
* docs/getting-started/setup/windows.md
  * Added alert box proposing a solution to issues regarding the
    installing the ffi gem on windows
* docs/getting-started/ssh.md
  * Added some regarding PuTTY and getting the private key to work with
    PuTTY