Work around 4.2.0 bug on Windows
VirtualBox 4.2.0 on Windows has a critical bug where "showvminfo
--machinereadable" only outputs 5 lines of output no matter what.
Vagrant relies heavily on this output to read things such as network
interfaces, shared folders, etc. Therefore, prior to this commit,
Vagrant was broken on Windows. Note that this is 100% a VIRTUALBOX BUG.
But enough users will use VirtualBox 4.2.0 on Windows that Vagrant must
work around it.

We work around it by introducing "jank mode" for the 4.2 driver on
Windows. On 4.2.0 only on Windows, Vagrant will use the human-readable
version of "showvminfo" and use some pretty crazy regular expressions to
parse out the information it needs. This approach is far less tested but
it at least gets things working again on Windows with 4.2.0. I'm hoping
in the future that Oracle will quickly fix this issue for a 4.2.1, but
I'm not certain yet.

Even if Oracle fixes the issue, this workaround will be in Vagrant
forever to accomodate legacy 4.2.0 users on Windows.