Add ask_secret to Vagrant::UI::Basic
Ruby unfortunately have a standard method for prompting for passwords without displaying the output. The methodologies change between UNIX platforms with `stty` available, Windows, and jRuby, and handling all cases can get complicated. https://gist.github.com/2040373 gives an example of how to cope for other platforms. The 'highline' gem supports doing this, but is an extra dependency that may be undesired.

Due to these complications, I've stuck to just using the `stty` approach. If it fails, it will fallback to a standard `#ask`, but warning the user that their input will be displayed.

`#ask` already does not appear jRuby compatible, which fails the `$stdin.tty?` test.