Defining the group of a shared folder
Symtom: The group ID for groups that are not backed by a user account (e.g. admin) cannot be obtained from the id command.

Consequence: The shared folder group can only be set to a group where by a user account of the same name exists.

Solution: Correcting the method to define the group ID by looking up the group ID via the group name.