Updates Chef provisioner configuration option formatter for Chef 11
Refs #1250.

Will correctly produce the desired result on Chef 11.x and above, as the
original directive was written against Chef 10.x, and released in Vagrant 1.2.7.

While this will continue to work for Chef 10.x, since this is not the 'mainline'
release track, and the alternative would be to write more conditional code in
the configuration file to detect the version and place the correct directive, this
changes the directive to support the current releases of Chef 11 and above.