persistent guest mount for linux
After the change where the provisioning is only run the first 'boot' or
forced we have encoutered some issues of services not starting because
of the files/folders were missing due to the shared folders to the host
were not yet there during startup.

Example:
- website folder shared with guest
- guest starts
- apache starts, website folder does not exist, apache fails
- no provisioning, so after mounting the shares apache is still down
- result: we have a 'non-working' system

This is only a patch for linux guests, but it solves the described
problem.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>