synced_folders/nfs_guest: Guest export NFS shares
Added preliminary Guest NFS exporting. The guest exports the mount, the
host mounts it. It's just the opposite way around to normal Vagrant NFS
synced folders.

This allows watchers and other applications running on the guest to
function that don't work across NFS. It also helps remove issues we've
had with NFS client caching causing edits to not be available on the
guests. Plus no 'sudo' prompt is needed as the mount (well on OSX) can
be done without root access.