WIP: a first shot at Vagrant.require_plugin('bar', '1.0')
So, I added a version parameter to require_plugin, hoping to add support to load/enforce
a specific version of a plugin when this is used in a Vagrantfile.

Right now struggling to write a proper unit or acceptance test for it since I seem to be
unable to install plugins on the fly or within a bundler environment.

Feedback welcome.