Allow inventory to be a directory
Ansible's inventory file has two meanings.  It can either
be a file or a directory containing inventory file(s) [1].
Updated Vagrant's inventory_path config option to support
this meaning.

This is useful when wanting the generated inventory files,
placed into an alternate directory.  Ansible will fail if
the files in an inventory path are not properly formatted.

    datarock:tribe(inventory-rework)☭% ansible -i . all --list-hosts
    ERROR: Invalid ini entry: License - need more than 1 value to unpack

This helps to organize things such that ansible can be run
outside of `vagrant provision` to run additional playbooks
(e.g. for integration tests).

[1] https://github.com/ansible/ansible/pull/2230