Synced folders SMB ACL Fixes
SMB shared folders works by creating a shared folder for an user account and by mounting it in the remote VM using the same user accounts credentials.
Hence user credentials has to be passed both during enabling the share and providing the share.
While enabling the shared folders in the host all files (recursive) within the host has to shared for an user account with a proper ACL and only then the guest will have the permission to access the shared folder.
I added a fix by calling a method to check if SMB credentials are required
and caching the instance of SMB implementer class so that when the same is
called during providing the share, user need not type the same credentials
again.