Fix NFS exports being broken yet again on Linux
The Ruby Regexp.escape method does not escape forward slashes because
it is not a meta-character. It is unique to sed expressions. An
alternative delimiter may be specified by prefixing the address
expression with a backslash with the delimiter character folloing.

Use control character (ASCII code 0x1a) as expression delimiter so it is
very unlikely an identifier will conflict.