eliminate slightly confusing wording
It doesn't make sense to use the present tense when saying box files are
split into two components, when one of those components is the box file
and one is something else.

One better way of phrasing it would be to use the past tense: "Box files
were split into two separate components" but even that's not completely
correct, because the old format did not include the metadata JSON
document which is one of the new components.

So it's safer to just say that today there are two different components.