reverse order of two box components
The metadata is optional whereas the box file is required, so it makes
sense to list the box file first.  It's also easier on the reader's
brain to start with the more obvious and easily understandable item
(which they'll probably be expecting to read about anyway) and save
the more surprising item till later.