tell the user when we are checksumming a downloaded box
There was already a locale string:

    Calculating and comparing box checksum...

but it wasn't being used.  This meant that the UI appeared to hang for
no obvious reason after box download completed, and this could be for
quite a long time for big boxes.  Now the user gets informed as soon as
checksumming begins, and the final `vagrant box add` output looks like
this:

    ==> box: Loading metadata for box 'mybox.json'
        box: URL: file:///home/adam/path/to/mybox/catalog.json
    ==> box: Adding box 'mycorp/mybox' (v0.0.1) for provider: virtualbox
        box: Downloading: mybox.box
        box: Calculating and comparing box checksum...
    ==> box: Successfully added box 'mycorp/mybox' (v0.0.1) for 'virtualbox'!