correction install nfs client on RHEL
On Centos/Redhat just install nfs-utils is not enought. We need to start
rpcbind, else we have this error :

mount.nfs: rpc.statd is not running but is required for remote locking.
mount.nfs: Either use '-o nolock' to keep locks local, or start statd.
mount.nfs: an incorrect mount option was specified