Explain that older Vagrant plugin licenses may not work with new VMware
This is the result of an unhappy customer who raised a very valid point. If you
are not on the Vagrant mailing list, you would be unaware that upgrading VMware
would cause your plugin to stop functioning.

While this does not solve the problem of dispersing that information, it helps
clarify that plugin version support is for the latest version of the Vagrant
plugin and does not extend to previous ones.