make metadata file detection work on any protocol
The metadata_url? method is refactored a lot. Now the metadata file is
always downloaded entirely (with a max size of METADATA_SIZE_LIMIT) if
it is a remote file. Then the temporary or local file is opened and
parsed as if it was a real json metadata file. If everything went fine,
then it is indeed a metadata file. The temporary downloaded file then
deleted.