don't download metadata twice
Instead of returning true and re-downloading and parsing the same file
in the add_from_metadata method, return the parsed metadata and use it
directly.

Also, throw a BoxAddShortNotFound if the metadata file couldn't be found
in the metadata_url? method. Before, it was done when re-downloading the
metadata file.